取消因为接收心跳包导致不能正确接收到飞控数据的判断，更改帧尾的排序，改进了结构体的内存数据对齐需求

+++ src/Application.hpp
@@ -114,8 +114,8 @@ private:
     // 位置信息
     double _latitude = 0; // 纬度
     double _longitude = 0; // 经度
-    float _altitude = 0; // 绝对高度
-    float _relativeAltitude = 0; // 相对高度
+    double _altitude = 0; // 绝对高度
+    double _relativeAltitude = 0; // 相对高度
     // 姿态信息
     float _pitch = 0; // 俯仰
     float _roll = 0;  // 横滚
